+++ src/app/competition/competition.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ApiService } from '../api.service';
+import { StatisticsService } from '../statistics.service';
 
 @Component({
   selector: 'app-competition',
@@ -8,10 +9,16 @@ import { ApiService } from '../api.service';
 })
 export class CompetitionComponent implements OnInit {
   public api: ApiService;
+  public statistics: StatisticsService;
   public saving: boolean = false;
 
-  constructor(api: ApiService) {
+  constructor(api: ApiService, statistics: StatisticsService) {
     this.api = api;
+    this.statistics = statistics;
+  }
+
+  generate_scramble(): void {
+    this.scramble = this.statistics.generateScramble();
   }
 
   ngOnInit(): void {
@@ -23,38 +30,14 @@ export class CompetitionComponent implements OnInit {
   public stoppedDuration: any = 0;
   public started = null;
   public running = false;
-  public blankTime = '00:00.000';
-  public time = '00:00.000';
+  public blankTime = '00:00.00';
+  public time = '00:00.00';
   public round = 1;
   public competition = true;
   public ms = 0;
   public timeElapsed: any = null;
   public result = [];
   public scramble = '';
-
-  generate_scramble = (): void => {
-    const moves = [
-      ['U', "U'", 'U2'],
-      ['D', "D'", 'D2'],
-      ['R', "R'", 'R2'],
-      ['L', "L'", 'L2'],
-      ['F', "F'", 'F2'],
-      ['B', "B'", 'B2'],
-    ];
-
-    let scramble = '';
-    let lastMoveType = null;
-
-    for (let i = 0; i < 20; i++) {
-      let moveType = Math.floor(Math.random() * 6);
-      moveType = moveType === lastMoveType ? (moveType + 1) % 6 : moveType;
-      const move = Math.floor(Math.random() * 3);
-      scramble += ` ${moves[moveType][move]}`;
-      lastMoveType = moveType;
-    }
-
-    this.scramble = scramble.trim();
-  };
 
   start() {
     if (this.running) return;
@@ -115,19 +98,11 @@ export class CompetitionComponent implements OnInit {
     this.timeElapsed = new Date(
       currentTime - this.timeBegan - this.stoppedDuration
     );
-    let min = this.timeElapsed.getUTCMinutes();
-    let sec = this.timeElapsed.getUTCSeconds();
-    let ms = this.timeElapsed.getUTCMilliseconds();
-    this.time =
-      this.zeroPrefix(min, 2) +
-      ':' +
-      this.zeroPrefix(sec, 2) +
-      '.' +
-      this.zeroPrefix(ms, 3);
     this.ms =
       this.timeElapsed.getUTCMilliseconds() +
       this.timeElapsed.getUTCSeconds() * 1000 +
       this.timeElapsed.getUTCMinutes() * 1000 * 60;
+    this.time = this.statistics.formatTime(this.ms)
   }
   async saveCompetition() {
     try {
